requirements.txt
numpy>=1.24.0,<2.0.0
pandas>=1.5.0,<2.0.0
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
ta>=0.10.0
yfinance>=0.2.30
Flask>=2.0.0
gunicorn>=21.0.0
Werkzeug>=2.0.0
requests>=2.28.0